requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.9
django-crispy-forms==1.13.0
django-debug-toolbar==3.2.2
django-heroku==0.3.1
django-timezone-field==4.2.1
Pillow==8.4.0
psycopg2==2.9.1
pytz==2021.3
sqlparse==0.4.2
whitenoise==5.3.0
